refactor: rename to dashboard-state-command
