+++ src/shared/commands/cell-footer-update-command.ts
@@ -1,5 +1,5 @@
 import { CellNotFoundError } from "../dashboard-state/dashboard-error";
-import DashboardStateCommand from "../dashboard-state/table-state-command";
+import DashboardStateCommand from "../dashboard-state/dashboard-state-command";
 import { FooterCell, DashboardState } from "../types";
 
 export default class CellFooterUpdateCommand extends DashboardStateCommand {
