+++ src/shared/commands/cell-header-update-command.ts
@@ -1,5 +1,5 @@
 import { CellNotFoundError } from "../dashboard-state/dashboard-error";
-import DashboardStateCommand from "../dashboard-state/table-state-command";
+import DashboardStateCommand from "../dashboard-state/dashboard-state-command";
 import { HeaderCell, DashboardState } from "../types";
 
 export default class CellHeaderUpdateCommand extends DashboardStateCommand {
