+++ src/shared/commands/row-sort-command.ts
@@ -4,7 +4,7 @@ import {
 	RowNotFoundError,
 	TagNotFoundError,
 } from "../dashboard-state/dashboard-error";
-import DashboardStateCommand from "../dashboard-state/table-state-command";
+import DashboardStateCommand from "../dashboard-state/dashboard-state-command";
 import {
 	BodyCell,
 	BodyRow,
