+++ src/shared/commands/tag-delete-command.ts
@@ -1,4 +1,4 @@
-import DashboardStateCommand from "../dashboard-state/table-state-command";
+import DashboardStateCommand from "../dashboard-state/dashboard-state-command";
 import { DashboardState, Tag } from "../types";
 import { TagNotFoundError } from "../dashboard-state/dashboard-error";
