+++ src/shared/dashboard-state/dashboard-state-command.ts
